core: add Symlink entry type
Add special entry type mimicing similar to symlinks on file systems:

The target entry is fetched from parent dict when actual method calls are done,
so it always reflects the current state in the parent dict.

This is useful, when entries shall be linked from one dict to another (or a
separate path inside the same), while always reflecting the target's current state.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>